pyproject.toml
[project]
name = "mcp_template_anthropic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.12",
    "langchain-mcp-adapters>=0.1.4",
    "langgraph>=0.4.1",
    "mcp>=1.9.2",
    "python-dotenv>=1.1.0",
]

[tool.setuptools]
packages = ["mcp_template"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"